backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
sse-starlette
pydantic
python-dotenv
pytest
black
isort
